Refactor: 커스텀 앱 바 뒤로 가기 버튼 있는 버전
